+++ flask_unchained/di.py
@@ -1,3 +1,6 @@
+import functools
+import inspect
+
 from py_meta_utils import (AbstractMetaOption, McsArgs, MetaOptionsFactory,
                            process_factory_meta_options, deep_getattr)
 from types import FunctionType
@@ -39,11 +42,39 @@ def ensure_service_name(service, name=None):
     return name
 
 
+def _inject_class_attrs(to_inject, fn=None):
+    def __init__(self, *args, **kwargs):
+        from .unchained import unchained
+        for param in to_inject:
+            try:
+                value = unchained.extensions.get(param, unchained.services[param])
+            except KeyError:
+                raise Exception(f'No service found with the name {param} '
+                                f'(required by {self.__class__.__name__})')
+            setattr(self, param, value)
+        if fn is not None:
+            return fn(self, *args, **kwargs)
+
+    if fn is not None:
+        return functools.wraps(fn)(__init__)
+    return __init__
+
+
 def set_up_class_dependency_injection(mcs_args: McsArgs):
-    if '__init__' in mcs_args.clsdict:
+    cls_attrs_to_inject = [k for k, v in mcs_args.clsdict.items() if v == injectable]
+    mcs_args.clsdict['__inject_cls_attrs__'] = cls_attrs_to_inject
+
+    if '__init__' not in mcs_args.clsdict:
+        init = _inject_class_attrs(cls_attrs_to_inject)
+        init.__di_name__ = mcs_args.name
+        init.__signature__ = inspect.signature(object)
+        mcs_args.clsdict['__init__'] = init
+        mcs_args.clsdict['__signature__'] = init.__signature__
+    else:
         from .unchained import unchained
         init = unchained.inject()(mcs_args.clsdict['__init__'])
-        mcs_args.clsdict['__init__'] = init
+        init.__di_name__ = mcs_args.name
+        mcs_args.clsdict['__init__'] = _inject_class_attrs(cls_attrs_to_inject, init)
         mcs_args.clsdict['__signature__'] = init.__signature__
 
     for attr, value in mcs_args.clsdict.items():